Add linting and unit test execution under CI
