+++ probe/test_load_config.py
@@ -13,7 +13,7 @@ class TestConfig(unittest.TestCase):
 
     def test_good_config_loads_successfully(self):
         """Validates proper loading of known good config"""
-        conf, _ = load_config("../testdata/test_good3.toml")
+        conf, _ = load_config("testdata/test_good3.toml")
         self.assertEqual(
             conf,
             {
@@ -78,19 +78,19 @@ class TestConfig(unittest.TestCase):
                 passwd="invalid",
             ),
         ]
-        config, msg = load_config("../testdata/test_good1.toml")
+        config, msg = load_config("testdata/test_good1.toml")
         self.assertIsNone(msg)
         actual_list = config_to_si_list(config)
         self.assertEqual(expect_list, actual_list)
 
     def test_display_parsed_config(self):
         """Validates that parsed config displayed on stderr matches expectations"""
-        config, msg = load_config("../testdata/test_good3.toml")
+        config, msg = load_config("testdata/test_good3.toml")
         self.assertIsNone(msg)
         mock_stderr = io.StringIO()
         display_parsed_config(config, file=mock_stderr)
         actual_output = mock_stderr.getvalue()
 
-        with open("../testdata/test_good3_parsed_display.txt", "rt") as expected:
+        with open("testdata/test_good3_parsed_display.txt", "rt") as expected:
             expected_output = expected.read()
             self.assertMultiLineEqual(actual_output, expected_output)
